# skills/gcms-content-assistant/references/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GCMS Automation API",
    "version": "1.0.0",
    "description": "Automation API for enabled languages, media uploads, post categories, link categories, post and link draft previews, posts, pages, and links. Replace the server URL with GCMS_API_BASE for the target site."
  },
  "servers": [
    {
      "url": "https://example.com/api/admin/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/page-design-context": {
      "get": {
        "operationId": "pages.design_context.get",
        "summary": "Read the live Pilot page-design context",
        "description": "Returns the current site identity, Hero, theme inheritance defaults, navigation, component registry, real data sources, page recipes, workflow policy, and required preview sizes.",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Versioned live page-design context" }
        }
      }
    },
...
    "/page-data-sources": {
      "get": {
        "operationId": "pages.data_sources.list",
        "summary": "List live GCMS data sources and binding fields",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Available live data sources" }
        }
      }
    },
    "/page-bindings/preview": {
      "post": {
        "operationId": "pages.bindings.preview",
        "summary": "Preview a manifest or component binding against live data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Resolved preview data without writing a revision" }
        }
      }
    },
...
    "/page-projects/{project_id}/capabilities/request": {
      "post": {
        "operationId": "pages.app.capabilities.request",
        "summary": "Request a manifest-declared app capability",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" },
          { "$ref": "#/components/parameters/PageIdempotencyKey" }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/PageCapabilityRequest"
        },
        "responses": {
          "201": { "description": "Capability request recorded" },
          "422": { "description": "Capability is undeclared or unavailable" }
        }
      }
    },
...
    "/page-projects/{project_id}/preview-url": {
      "post": {
        "operationId": "pages.preview",
        "summary": "Create a short-lived private preview for an exact revision/build",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" }
        ],
        "responses": { "201": { "description": "Signed no-store/noindex preview URL" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GCMS Access Token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-GCMS-API-Key"
      }
    },
...
    "parameters": {
      "PageProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 }
      },
...
      "PageIfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Exact ETag from the latest project response. Never guess or overwrite after a conflict.",
        "schema": { "type": "string" }
      },
      "PageIdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Stable only for byte-equivalent retries of one logical request.",
        "schema": { "type": "string", "minLength": 8, "maxLength": 200 }
      },
...
    },
...
    "requestBodies": {
      "PageCapabilityRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["capability"],
              "properties": {
                "capability": { "type": "string" },
                "config": { "type": "object", "additionalProperties": true }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
    },
...
  }
}